Refactoring unused code.
